refactor: Outsource common functionality of qrisp exact- & sat variants
